TRestStringHelper fixing compilation issue

+++ source/framework/tools/src/TRestStringHelper.cxx
@@ -119,7 +119,7 @@ string REST_StringHelper::ReplaceMathematicalExpressions(string buffer, Int_t pr
     }
 
     // we split the unit part and the expression part
-    int pos = buffer.find_first_of("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz ");
+    size_t pos = buffer.find_first_of("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz ");
 
 	if( pos == string::npos ) pos = 0;
